Command renamed to commands
