+++ src/Game/Commands/Command.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace BinaryStudioAcademy\Game\Commands;
+
+
+abstract class Command implements \BinaryStudioAcademy\Game\Contracts\Commands\Command
+{
+    public abstract function execute(string $params = '');
+}
