+++ src/Game/Commands/ExitCommand.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace BinaryStudioAcademy\Game\Command;
+namespace BinaryStudioAcademy\Game\Commands;
 
 use BinaryStudioAcademy\Game\Contracts\Io\Writer;
 
-class SchemeCommand extends Command
+class ExitCommand extends Command
 {
     private $writer;
 
@@ -15,6 +15,7 @@ class SchemeCommand extends Command
 
     public function execute(string $params = '')
     {
-        // TODO: Implement execute() method.
+        $this->writer->writeln('You exited game.');
+        exit;
     }
 }
