+++ src/Game/Commands/ProduceCommand.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace BinaryStudioAcademy\Game\Command;
+namespace BinaryStudioAcademy\Game\Commands;
 
 use BinaryStudioAcademy\Game\Contracts\Io\Writer;
